pyproject.toml
[tool.poetry]
name = "proxypool"
version = "0.1.0"
description = ""
authors = ["sube <z397503810@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
hydra-core = "^1.3.2"
environs = "^9.5.0"
loguru = "^0.7.0"
pydantic = "^2.3.0"
redis = "^5.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
